Updated comments in main.py and a new file named database.py
